correcciones controlador
