+++ controllers/authors/update.js
@@ -5,6 +5,7 @@ const controller = {
     update: async(req,res,next) => {
         try{
             let author = await Author.findOneAndUpdate({ user_id: req.user }, req.body, { new: true })
+            .select("name last_name city country date photo")
             if(author){
                 return res.status(200).json({
                     success: true,
@@ -12,7 +13,6 @@ const controller = {
                 })
             }
             return next ( createError(404, "no es author" ))
-
         }catch(error){
             return res.status(400).json({
                 success: false,
